Some packaging and CHANGELOG  correction.
