+++ kiel/src/main/java/me/ibrahimyilmaz/kiel/adapter/OnViewHolderBound.kt
@@ -1,0 +1,3 @@
+package me.ibrahimyilmaz.kiel.adapter
+
+typealias OnViewHolderBound<T, VH> = (VH, Int, T) -> Unit
