+++ kiel/src/main/java/me/ibrahimyilmaz/kiel/adapter/OnViewHolderBoundWithPayload.kt
@@ -1,0 +1,3 @@
+package me.ibrahimyilmaz.kiel.adapter
+
+typealias OnViewHolderBoundWithPayload<T, VH> = (VH, Int, T, List<Any>) -> Unit
