+++ kiel/src/main/java/me/ibrahimyilmaz/kiel/adapter/OnViewHolderCreated.kt
@@ -1,0 +1,3 @@
+package me.ibrahimyilmaz.kiel.adapter
+
+typealias OnViewHolderCreated<VH> = (VH) -> Unit
