+++ samples/src/main/java/me/ibrahimyilmaz/kiel/samples/adapter/viewholder/TextMessageViewHolder.kt
@@ -1,10 +1,10 @@
-package me.ibrahimyilmaz.kiel.samples.datasource.viewbinder
+package me.ibrahimyilmaz.kiel.samples.adapter.viewholder
 
 import android.view.View
 import android.widget.TextView
 import me.ibrahimyilmaz.kiel.adapter.RecyclerViewHolder
 import me.ibrahimyilmaz.kiel.samples.R
-import me.ibrahimyilmaz.kiel.samples.datasource.model.MessageViewState.Text
+import me.ibrahimyilmaz.kiel.samples.adapter.model.MessageViewState.Text
 
 
 class TextMessageViewHolder(view: View) : RecyclerViewHolder<Text>(view) {
