+++ samples/src/main/res/values/strings.xml
@@ -1,4 +1,4 @@
 <resources>
     <string name="app_name">Kiel-Samples</string>
-    <string name="fragment_datasource_example">Environments</string>
+    <string name="fragment_recyclerviewadapter_example">RecyclerViewAdapter</string>
 </resources>
